requirements.txt
# requirements.txt
python-telegram-bot==13.15
python-dotenv==0.21.1
openai==0.27.0
firebase-admin==6.1.0  # If using Firebase
pymongo==4.3.3  # If using MongoDB
apscheduler==3.6.3  # For scheduling tasks